requirements.txt
charset-normalizer==3.4.2
pillow==11.3.0
reportlab==4.4.2
